refactor: Rename package from ovnexporter to ovn-exporter
